osint_tool/requirements.txt
telethon
python-dotenv
colorama